% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dPiecewisePareto}
\alias{dPiecewisePareto}
\title{Density of the Piecewise Pareto Distribution}
\usage{
dPiecewisePareto(x, t, alpha, truncation = NULL, truncation_type = "lp")
}
\arguments{
\item{x}{Numeric. The function evaluates the density at \code{x}.}

\item{t}{Numeric vector. Thresholds of the piecewise Pareto distribution.}

\item{alpha}{Numeric vector. \code{alpha[i]} is the Pareto alpha in excess of \code{t[i]}.}

\item{truncation}{Numeric. If truncation is not NULL and truncation > t, then the Pareto distribution is truncated at truncation.}

\item{truncation_type}{Character. If \code{truncation_type = "wd"} then the whole distribution is truncated. If \code{truncation_type = "lp"} then a truncated Pareto is used for the last piece.}
}
\value{
Density function of the piecewise Pareto distribution with parameter vectors \code{t} and \code{alpha} evaluated at \code{x}
}
\description{
Calculates the density function of the piecewise Pareto distribution
}
\examples{
t <- c(1000, 2000, 3000)
alpha <- c(1, 1.5, 2)
x <- 0:10 * 1000
dPiecewisePareto(x, t, alpha)
dPiecewisePareto(x, t, alpha, truncation = 5000, truncation_type = "lp")
dPiecewisePareto(x, t, alpha, truncation = 5000, truncation_type = "wd")

}
