% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R
\docType{data}
\name{Example1_AP}
\alias{Example1_AP}
\title{Example data: Attachment Points}
\format{An object of class \code{numeric} of length 5.}
\usage{
Example1_AP
}
\description{
Example data: Attachment Points
}
\keyword{datasets}
