% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R
\docType{data}
\name{Example1_EL}
\alias{Example1_EL}
\title{Example data: Expected Losses}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
Example1_EL
}
\description{
Example data: Expected Losses
}
\keyword{datasets}
