% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectiveModelMethods.R
\name{Excess_Frequency.PGP_Model}
\alias{Excess_Frequency.PGP_Model}
\title{Expected Frequency in Excess of a Threshold}
\usage{
\method{Excess_Frequency}{PGP_Model}(CollectiveModel, x = 0)
}
\arguments{
\item{CollectiveModel}{PGP_Model object.}

\item{x}{Numeric. Threshold.}
}
\value{
The expected frequency in excess of \code{x} for the given \code{CollectiveModel}
}
\description{
Calculates the expected frequency in excess of a threshold for a PGP_model
}
\examples{
PGPM <- PGP_Model(2, 1000, 1, 2, dispersion = 2)
PGPM
Excess_Frequency(PGPM, c(-Inf, 0, 1000, 2000, 3000, Inf))

}
