% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectiveModelMethods.R
\name{Excess_Frequency.PPP_Model}
\alias{Excess_Frequency.PPP_Model}
\title{Expected Frequency in Excess of a Threshold}
\usage{
\method{Excess_Frequency}{PPP_Model}(CollectiveModel, x = 0)
}
\arguments{
\item{CollectiveModel}{PPP_Model object.}

\item{x}{Numeric. Threshold.}
}
\value{
The expected frequency in excess of \code{x} for the given \code{CollectiveModel}
}
\description{
Calculates the expected frequency in excess of a threshold for a PPP_model
}
\examples{
PPPM <- PiecewisePareto_Match_Layer_Losses(Example1_AP, Example1_EL)
PPPM
Excess_Frequency(PPPM, c(-Inf, 0, 1000, 2000, 3000, Inf))

}
