\name{inner2Inter}
\alias{inner2Inter}
\title{Inner vs. inter Cluster distances}
\description{
 This function calculates the expected Ratio of Inner vs. inter Cluster distances.}
\usage{
inner2Inter(expectedNumberOfClusters)
}
\arguments{
  \item{expectedNumberOfClusters}{numeric value, the number of clusters that are expected in a data set. 6 by default.}
}

\value{
  numeric value
}
\references{Matlab dbt - Pareto}
\author{Friederike Matz}

