% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGPModel.R
\name{is.PGP_Model}
\alias{is.PGP_Model}
\title{Check if an object is a PGP_Model}
\usage{
is.PGP_Model(x)
}
\arguments{
\item{x}{Object to be checked.}
}
\description{
Checks if the class of an object is 'PGP_Model'
}
\examples{
PGPM <- PGP_Model(2, 1000, 1, 2, dispersion = 2)
PGPM
is.valid.PGP_Model(PGPM)
is.valid.PGP_Model(PGPM, comment = TRUE)

PGPM$alpha_tail <- -2
is.PGP_Model(PGPM)
is.valid.PGP_Model(PGPM)
is.valid.PGP_Model(PGPM, comment = TRUE)

}
