% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPPModel.R
\name{is.PPP_Model}
\alias{is.PPP_Model}
\title{Check if an object is a PPP_Model}
\usage{
is.PPP_Model(x)
}
\arguments{
\item{x}{Object to be checked.}
}
\description{
Checks if the class of an object is 'PPP_Model'
}
\examples{
PPPM <- PPP_Model(2, c(1000,2000), c(1,2), dispersion = 2)
PPPM
is.valid.PPP_Model(PPPM)

PPPM$alpha <- 2
is.valid.PPP_Model(PPPM)
is.PPP_Model(PPPM)

}
