\name{LKWFahrzeitSeehafen2010}
\alias{LKWFahrzeitSeehafen2010}
\docType{data}
\title{
LKWFahrzeitSeehafen2010
}
\description{
...~
}
\usage{data("LKWFahrzeitSeehafen2010")}
\format{
  The format is:
 num [1:11441] 84.7 13.2 11.5 41.4 52.9 ...
}
\details{
...
}
\source{
...
}
\references{
wida, behnisch
}
\examples{
data(LKWFahrzeitSeehafen2010)
## maybe str(LKWFahrzeitSeehafen2010) ; plot(LKWFahrzeitSeehafen2010) ...
}
\keyword{datasets}
