% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen.R
\name{Local_Pareto_Alpha}
\alias{Local_Pareto_Alpha}
\title{This function calculates the local Pareto alpha of a distribution}
\usage{
Local_Pareto_Alpha(x, distribution, ...)
}
\arguments{
\item{x}{Numeric. Vector of thresholds at which the local Pareto alpha is calculated.}

\item{distribution}{Character.
'lnorm' for lognormal distribution (parameters meanlog, sdlog)
'norm' for normal distribution (parameters: mean, sd)
"gamma' for gamma distribution (parameters: shape, rate, scale)}

\item{...}{Parameters for distribution}
}
\description{
This function calculates the local Pareto alpha of a distribution
}
