\name{pareto-package}
\alias{pareto-package}
\alias{pareto}
\docType{package}
\title{
\packageTitle{pareto}
}
\description{
\packageDescription{pareto}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pareto}
\packageIndices{pareto}
Density, distribution function, quantile function and random
  generation for the pareto distribution with parameters \code{alpha} and
  \code{beta}.
}
\author{
\packageAuthor{pareto}

Maintainer: \packageMaintainer{pareto}
}
\keyword{ package }


