% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen.R
\name{Pareto_CDF}
\alias{Pareto_CDF}
\title{This function calculates the cumulative distribution function of a Pareto Distribution}
\usage{
Pareto_CDF(x, t, alpha, truncation = NULL)
}
\arguments{
\item{x}{Numeric. The function evaluates the CDF at x.}

\item{t}{Numeric. Threshold of the piecewise Pareto distribution.}

\item{alpha}{Numeric. Pareto alpha.}

\item{truncation}{Numeric. If truncation is not NULL and truncation > t, then the Pareto distribution is truncated at truncation.}
}
\description{
This function calculates the cumulative distribution function of a Pareto Distribution
}
