% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Pareto_Extrapolation}
\alias{Pareto_Extrapolation}
\title{This function extrapolates from the layer Cover_1 xs AttachmentPoint_1 to the layer Cover_2 xs AttachmentPoint_2 using a Pareto distribution Pareto(t, alpha) with t sufficiently small.}
\usage{
Pareto_Extrapolation(
  Cover_1,
  AttachmentPoint_1,
  Cover_2,
  AttachmentPoint_2,
  alpha,
  ExpLoss_1 = NULL,
  truncation = NULL
)
}
\arguments{
\item{Cover_1}{Numeric. Cover of the layer from which we extrapolate. Use Inf for unlimited layers.}

\item{AttachmentPoint_1}{Numeric. Attachment point of the layer from which we extrapolate.}

\item{Cover_2}{Numeric. Cover of the layer to which we extrapolate. Use Inf for unlimited layers.}

\item{AttachmentPoint_2}{Numeric. Attachment point of the layer to which we extrapolate.}

\item{alpha}{Numeric. Pareto alpha used for the extrapolation.}

\item{ExpLoss_1}{Numeric. Expected loss of the layer from which we extrapolate. If NULL (default) then the function provides only the ratio between the expected losses of the layers.}

\item{truncation}{Numeric. If truncation is not NULL and truncation > AttachmentPoint_1, then the Pareto distribution is truncated at truncation.}
}
\description{
This function extrapolates from the layer Cover_1 xs AttachmentPoint_1 to the layer Cover_2 xs AttachmentPoint_2 using a Pareto distribution Pareto(t, alpha) with t sufficiently small.
}
