% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen.R
\name{Pareto_Find_Alpha_btw_FQ_Layer}
\alias{Pareto_Find_Alpha_btw_FQ_Layer}
\title{This function finds the Pareto alpha between an excess frequency and a layer.}
\usage{
Pareto_Find_Alpha_btw_FQ_Layer(Threshold, Frequency, Cover,
  AttachmentPoint, ExpLoss, max_alpha = 100, tolerance = 10^(-10),
  truncation = NULL)
}
\arguments{
\item{Threshold}{Numeric. Threshold}

\item{Frequency}{Numeric. Expected frequency in excess of Thershold}

\item{Cover}{Numeric. Cover of the second layer.}

\item{AttachmentPoint}{Numeric. Cover of the layer.}

\item{ExpLoss}{Numeric. Expected loss of the layer.}

\item{max_alpha}{Numeric. Upper limit for the alpha that is returned.}

\item{tolerance}{Numeric. Accuracy of the result.}

\item{truncation}{Numeric. If truncation is not NULL then the Pareto distribution is truncated at truncation.}
}
\description{
This function finds the Pareto alpha between an excess frequency and a layer.
}
