% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Pareto_Find_Alpha_btw_FQs}
\alias{Pareto_Find_Alpha_btw_FQs}
\title{This function finds the Pareto alpha between two excess frequencies.}
\usage{
Pareto_Find_Alpha_btw_FQs(
  Threshold_1,
  Frequency_1,
  Threshold_2,
  Frequency_2,
  max_alpha = 100,
  tolerance = 1e-10,
  truncation = NULL
)
}
\arguments{
\item{Threshold_1}{Numeric. Threshold 1}

\item{Frequency_1}{Numeric. Expected frequency in excess of Thershold 1}

\item{Threshold_2}{Numeric. Threshold 2}

\item{Frequency_2}{Numeric. Expected frequency in excess of Thershold 2}

\item{max_alpha}{Numeric. Upper limit for the alpha that is returned.}

\item{tolerance}{Numeric. Accuracy of the result.}

\item{truncation}{Numeric. If truncation is not NULL then the Pareto distribution is truncated at truncation.}
}
\description{
This function finds the Pareto alpha between two excess frequencies.
}
