% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen.R
\name{Pareto_Find_Alpha_btw_Layers}
\alias{Pareto_Find_Alpha_btw_Layers}
\title{This function finds the Pareto alpha between two layers.}
\usage{
Pareto_Find_Alpha_btw_Layers(Cover_1, AttachmentPoint_1, ExpLoss_1, Cover_2,
  AttachmentPoint_2, ExpLoss_2, max_alpha = 20, tolerance = 10^(-10))
}
\arguments{
\item{Cover_1}{Numeric. Cover of the first layer.}

\item{AttachmentPoint_1}{Numeric. Cover of the first layer.}

\item{ExpLoss_1}{Numeric. Expected loss of the first layer.}

\item{Cover_2}{Numeric. Cover of the second layer.}

\item{AttachmentPoint_2}{Numeric. Cover of the second layer.}

\item{ExpLoss_2}{Numeric. Expected loss of the second layer.}

\item{max_alpha}{Numeric. Upper limit for the alpha that is returned.}

\item{tolerance}{Numeric. Accuracy of the result.}
}
\description{
This function finds the Pareto alpha between two layers.
}
