% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Pareto_Inverse_CDF}
\alias{Pareto_Inverse_CDF}
\title{This function calculates the inverse of the cumulative distribution function of a Pareto Distribution}
\usage{
Pareto_Inverse_CDF(y, t, alpha, truncation = NULL)
}
\arguments{
\item{y}{Numeric. The function evaluates the inverse CDF at y.}

\item{t}{Numeric. Threshold of the piecewise Pareto distribution.}

\item{alpha}{Numeric. Pareto alpha.}

\item{truncation}{Numeric. If truncation is not NULL and truncation > t, then the Pareto distribution is truncated at truncation.}
}
\description{
This function calculates the inverse of the cumulative distribution function of a Pareto Distribution
}
