% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen.R
\name{Pareto_Layer_Mean}
\alias{Pareto_Layer_Mean}
\title{This function the expected loss of the Pareto Distribution Pareto(t, alpha) in the layer Cover xs AttachmentPoint}
\usage{
Pareto_Layer_Mean(Cover, AttachmentPoint, alpha, t = NULL)
}
\arguments{
\item{Cover}{Numeric. Cover of the reinsurance layer. Use "unl" for unlimited layers.}

\item{AttachmentPoint}{Numeric. Attachment point of the reinsurance layer.}

\item{alpha}{Numeric. Pareto alpha.}

\item{t}{Numeric.. Threshold of the Pareto distribution. If t = NULL (default) then t <- Attachment Point}
}
\description{
This function the expected loss of the Pareto Distribution Pareto(t, alpha) in the layer Cover xs AttachmentPoint
}
