% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen.R
\name{Pareto_Layer_SM}
\alias{Pareto_Layer_SM}
\title{This function the second moment of the xs-loss of the Pareto Distribution Pareto(t, alpha) in the layer Cover xs AttachmentPoint}
\usage{
Pareto_Layer_SM(Cover, AttachmentPoint, alpha, t = NULL, truncation = NULL)
}
\arguments{
\item{Cover}{Numeric. Cover of the reinsurance layer. Use Inf for unlimited layers.}

\item{AttachmentPoint}{Numeric. Attachment point of the reinsurance layer.}

\item{alpha}{Numeric. Pareto alpha.}

\item{t}{Numeric. Threshold of the Pareto distribution. If t = NULL (default) then t <- Attachment Point}

\item{truncation}{Numeric. If truncation is not NULL and truncation > t, then the Pareto distribution is truncated at truncation.}
}
\description{
This function the second moment of the xs-loss of the Pareto Distribution Pareto(t, alpha) in the layer Cover xs AttachmentPoint
}
