\name{PDEbins}
\alias{PDEbins}

\title{PDE Bins}
\description{
This function plots PDE with vertical lines for bins.}
\usage{
PDEbins(kernels,paretoDensity,bins,numbers=FALSE)
}

\arguments{
  \item{kernels}{numeric vector of kernels from PDE.}
  \item{paretoDensity}{numeric vector of density values at kernels.}
  \item{bins}{numeric vector of bin boundaries.}
  \item{numbers}{logical. Whether to show numbers. FALSE by default.}
}



\references{Matlab dbt - Pareto}
\author{Friederike Matz}