\name{PDEcls}
\alias{PDEcls}
\title{PDE class}
\description{
This function plots PDE densities for each class.}
\usage{
PDEcls(data, cls, show_legend=FALSE)
}

\arguments{
  \item{data}{numeric matrix of data to be plotted.}
  \item{cls}{numeric vector. class number for each dataset. }
  \item{show_legend}{logical. If to show legend. FALSE by default.}
}

\references{Matlab dbt - Pareto}
\author{Friederike Matz }
