% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen.R
\name{PiecewisePareto_CDF}
\alias{PiecewisePareto_CDF}
\title{This function calculates the cumulative distribution function of a Piecewise Pareto Distribution}
\usage{
PiecewisePareto_CDF(x, t, alpha)
}
\arguments{
\item{x}{Numeric. The function evaluates the CDF at x.}

\item{t}{Numeric vector. Thresholds of the piecewise Pareto distribution.}

\item{alpha}{Numeric vector. Pareto alpha[i] = Pareto alpha in excess of t[i].}
}
\description{
This function calculates the cumulative distribution function of a Piecewise Pareto Distribution
}
