% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen.R
\name{PiecewisePareto_Inverse_CDF}
\alias{PiecewisePareto_Inverse_CDF}
\title{This function calculates the inverse of the cumulative distribution function of a Piecewise Pareto Distribution}
\usage{
PiecewisePareto_Inverse_CDF(y, t, alpha, truncation = NULL,
  truncation_type = "lp")
}
\arguments{
\item{y}{Numeric. The function evaluates the inverse CDF at y.}

\item{t}{Numeric vector. Thresholds of the piecewise Pareto distribution.}

\item{alpha}{Numeric vector. Pareto alpha[i] = Pareto alpha in excess of t[i].}

\item{truncation}{Numeric. If truncation is not NULL and truncation > t, then the Pareto distribution is truncated at truncation.}

\item{truncation_type}{Charakter. If truncation_type = "wd" then the whole distribution is truncated. If truncation_type = "lp" then a truncated Pareto is used for the last piece.}
}
\description{
This function calculates the inverse of the cumulative distribution function of a Piecewise Pareto Distribution
}
