% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen.R
\name{PiecewisePareto_Layer_Mean}
\alias{PiecewisePareto_Layer_Mean}
\title{This function calculates the expected loss of a Piecewise Pareto Distribution in the layer Cover xs AttachmentPoint}
\usage{
PiecewisePareto_Layer_Mean(Cover, AttachmentPoint, t, alpha)
}
\arguments{
\item{Cover}{Numeric. Cover of the reinsurance layer.}

\item{AttachmentPoint}{Numeric. Attachment point of the reinsurance layer.}

\item{t}{Numeric vector. Thresholds of the piecewise Pareto distribution.}

\item{alpha}{Numeric vector. Pareto alpha[i] = Pareto alpha in excess of t[i].}
}
\description{
This function calculates the expected loss of a Piecewise Pareto Distribution in the layer Cover xs AttachmentPoint
}
