% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{pPareto}
\alias{pPareto}
\title{Distribution Function of the Pareto Distribution}
\usage{
pPareto(x, t, alpha, truncation = NULL)
}
\arguments{
\item{x}{Numeric. The function evaluates the CDF at x.}

\item{t}{Numeric. Threshold of the piecewise Pareto distribution.}

\item{alpha}{Numeric. Pareto alpha.}

\item{truncation}{Numeric. If truncation is not NULL and truncation > t, then the Pareto distribution is truncated at truncation.}
}
\value{
Distribution function of the Pareto distribution with parameters \code{t} and \code{alpha} evaluated at \code{x}
}
\description{
Calculates the cumulative distribution function of a Pareto distribution
}
\examples{
x <- 0:10 * 1000
pPareto(x, 1000, 2)
pPareto(x, 1000, 2, truncation = 5000)


}
