% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPPModel.R
\name{PPP_Model_Excess_Frequency}
\alias{PPP_Model_Excess_Frequency}
\title{Expected Frequency in Excess of a Threshold}
\usage{
PPP_Model_Excess_Frequency(x, PPP_Model)
}
\arguments{
\item{x}{Numeric. Threshold.}

\item{PPP_Model}{PPP_Model object.}
}
\value{
The expected frequency in excess of \code{x} for the given \code{PPP_Model}
}
\description{
Calculates the expected frequency in excess of a threshold for a PPP_Model
}
\examples{
PPPM <- PiecewisePareto_Match_Layer_Losses(Example1_AP, Example1_EL)
PPPM
Excess_Frequency(PPPM, c(-Inf, 0, 1000, 2000, 3000, Inf))

}
