% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPPModel.R
\name{PPP_Model_Simulate}
\alias{PPP_Model_Simulate}
\title{Simulate Losses with a PPP_Model}
\usage{
PPP_Model_Simulate(n, PPP_Model)
}
\arguments{
\item{n}{Integer. Number of Simulations.}

\item{PPP_Model}{PPP_Model object.}
}
\value{
A matrix where row k contains the simulated losses of the kth simulation.
}
\description{
Simulates losses of a PPP_Model
}
\examples{
PPPM <- PiecewisePareto_Match_Layer_Losses(c(1000, 2000, 3000), c(2000, 1000, 500),
                                           truncation = 10000, truncation_type = "wd")
PPPM
Simulate_Losses(PPPM, 100)

}
