% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGPModel.R
\name{print.PGP_Model}
\alias{print.PGP_Model}
\title{Print a PGP_Model Object(Collective Panjer & Generalized Pareto Model) Object}
\usage{
\method{print}{PGP_Model}(x, ...)
}
\arguments{
\item{x}{PGP_Model object.}

\item{...}{Other arguments, all currently ignored.}
}
\description{
Print method for PGP_Model objects
}
