% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPPModel.R
\name{print.PPP_Model}
\alias{print.PPP_Model}
\title{Print a PPP_Model Object(Collective Panjer & Piecewise Pareto Model) Object}
\usage{
\method{print}{PPP_Model}(x, ...)
}
\arguments{
\item{x}{PPP_Model object.}

\item{...}{Other arguments, all currently ignored.}
}
\description{
Print method for PPP_Model objects
}
