% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectiveModelMethods.R
\name{Simulate_Losses.PGP_Model}
\alias{Simulate_Losses.PGP_Model}
\title{Simulate Losses with a PGP_Model}
\usage{
\method{Simulate_Losses}{PGP_Model}(CollectiveModel, nyears = 1)
}
\arguments{
\item{CollectiveModel}{PGP_Model object.}

\item{nyears}{Integer. Number of simulated years.}
}
\value{
A matrix where row k contains the simulated losses of the kth simulated year.
}
\description{
Simulates losses with a PGP_Model
}
\examples{
PGPM <- PGP_Model(2, 1000, 1, 2, dispersion = 2)
PGPM
Simulate_Losses(PGPM, 100)

}
