% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Local_Pareto_Alpha}
\alias{Local_Pareto_Alpha}
\title{Local Pareto Alpha}
\usage{
Local_Pareto_Alpha(x, distribution, ...)
}
\arguments{
\item{x}{Numeric. Vector of thresholds at which the local Pareto alpha is calculated.}

\item{distribution}{Character. \itemize{
\item 'lnorm' for lognormal distribution (parameters meanlog, sdlog)
\item 'norm' for normal distribution (parameters: mean, sd)
\item 'gamma' for gamma distribution (parameters: shape, rate, scale)
}}

\item{...}{Parameters for the selected distribution}
}
\value{
Local Pareto alpha of the selected distribution at \code{x}
}
\description{
Calculates the local Pareto alpha of the normal, lognormal and gamma distribution
}
