\name{pareto-package}
\alias{pareto-package}
\alias{pareto}
\docType{package}
\title{
\packageTitle{pareto}
}
\description{
\packageDescription{pareto}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pareto}
\packageIndices{pareto}
 The package is created to compute the pdf/cdf/quantile of pareto distribution. the main functions are dpareto/ppareto/qpareto. Specially, function dpareto will call rcpp file "dpareto.cpp" to compute pareto density. 
}
\author{
\packageAuthor{pareto}

Maintainer: \packageMaintainer{pareto}
}
\keyword{ package }