\name{pareto-package}
\alias{pareto-package}
\alias{pareto}
\docType{package}
\title{
\packageTitle{pareto}
}
\description{
\packageDescription{pareto}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pareto}
\packageIndices{pareto}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{pareto}

Maintainer: \packageMaintainer{pareto}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
dpareto( 2:4, 1, 2)
}
