% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Pareto_ML_Estimator_Alpha}
\alias{Pareto_ML_Estimator_Alpha}
\title{This function calculates the maximum likelihood estimator of the parameter alpha of a Pareto distribution}
\usage{
Pareto_ML_Estimator_Alpha(
  losses,
  t,
  truncation = NULL,
  alpha_min = 0.001,
  alpha_max = 10
)
}
\arguments{
\item{losses}{Numeric vector. Losses for ML estimation.}

\item{t}{Numeric. Threshold of the Pareto distribution.}

\item{truncation}{Numeric. If truncation is not NULL and truncation > t, then the Pareto distribution is truncated at truncation.}

\item{alpha_min}{Numeric. Lower bound for alpha.}

\item{alpha_max}{Numeric. Upper bound for alpha.}
}
\description{
This function calculates the maximum likelihood estimator of the parameter alpha of a Pareto distribution
}
