% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen.R
\name{PiecewisePareto_Layer_Mean}
\alias{PiecewisePareto_Layer_Mean}
\title{This function calculates the expected loss of a Piecewise Pareto Distribution in the layer Cover xs AttachmentPoint}
\usage{
PiecewisePareto_Layer_Mean(Cover, AttachmentPoint, t, alpha,
  truncation = NULL, truncation_type = "lp")
}
\arguments{
\item{Cover}{Numeric. Cover of the reinsurance layer.}

\item{AttachmentPoint}{Numeric. Attachment point of the reinsurance layer.}

\item{t}{Numeric vector. Thresholds of the piecewise Pareto distribution.}

\item{alpha}{Numeric vector. Pareto alpha[i] = Pareto alpha in excess of t[i].}

\item{truncation}{Numeric. If truncation is not NULL and truncation > t, then the Pareto distribution is truncated at truncation.}

\item{truncation_type}{Charakter. If truncation_type = "wd" then the whole distribution is truncated. If truncation_type = "lp" then a truncated Pareto is used for the last piece.}
}
\description{
This function calculates the expected loss of a Piecewise Pareto Distribution in the layer Cover xs AttachmentPoint
}
