\name{ppareto}
\alias{ppareto}
\title{
compute cumulative distribution function of pareto distribution 
}
\description{
  it will compute cumulative distribution function of pareto distribution 
}
\usage{
ppareto(q = 2, alpha = 1, beta = 1, lower.tail = TRUE,
                    log.p = FALSE)
}
\arguments{
  \item{q}{
   vector q
}
  \item{alpha}{
   vector alpha
}
  \item{beta}{
   vector beta
}
  \item{lower.tail}{
   vector lower.tail
}
  \item{log.p}{
   vector log.p
}
}
\references{
  \href{https://en.wikipedia.org/wiki/Pareto_distribution}{Wikipedia pareto distribution}
}
\author{
 Ruifeng Chen
}
\keyword{ ppareto }% use one of  RShowDoc("KEYWORDS")
