\name{qpareto}
\alias{qpareto}
\title{
compute quantile of pareto distrition
}
\description{
  it will compute quantile of pareto distrition
}
\usage{
qpareto(p = 0.5, alpha = 1, beta = 1, lower.tail = TRUE,
                    log.p = FALSE)
}
\arguments{
  \item{p}{
   vector p
}
  \item{alpha}{
   vector alpha
}
  \item{beta}{
   vector beta
}
  \item{lower.tail}{
   vector lower.tail
}
  \item{log.p}{
   vector log.p
}
}
\references{
  \href{https://en.wikipedia.org/wiki/Pareto_distribution}{Wikipedia pareto distribution}
}
\author{
 Ruifeng Chen
}
\keyword{ qpareto }% use one of  RShowDoc("KEYWORDS")
