% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen.R
\name{rPareto}
\alias{rPareto}
\title{This function generates random deviates of the Pareto distribution Pareto(t, alpha).}
\usage{
rPareto(n, t, alpha, truncation = NULL)
}
\arguments{
\item{n}{Number of observations.}

\item{t}{Threshold of the Pareto distribution}

\item{alpha}{Pareto alpha.}

\item{truncation}{If truncation is not NULL and truncation > t, then the Pareto distribution is truncated at truncation (resampled Pareto)}
}
\description{
This function generates random deviates of the Pareto distribution Pareto(t, alpha).
}
