% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen.R
\name{rPiecewisePareto}
\alias{rPiecewisePareto}
\title{This function generates random deviates of the Piecewise Pareto distribution.}
\usage{
rPiecewisePareto(n, t, alpha)
}
\arguments{
\item{n}{Numeric. Number of simulations}

\item{t}{Numeric vector. Thresholds of the piecewise Pareto distribution.}

\item{alpha}{Numeric vector. Pareto alpha[i] = Pareto alpha in excess of t[i].}
}
\description{
This function generates random deviates of the Piecewise Pareto distribution.
}
